requirement.txt
Django==2.1.5
django-allauth==0.39.1
django-crispy-forms==1.7.2
djangorestframework==3.9.2
django-filter==2.1.0
requests==2.20.1
